+++ radio/src/CMakeLists.txt
@@ -473,6 +473,7 @@ set(SRC
   serial.cpp
   sbus.cpp
   hal/module_port.cpp
+  tasks/mixer_task.cpp
   )
 
 if(GUI)
