+++ radio/src/mixer_scheduler.cpp
@@ -21,6 +21,7 @@
 
 #include "opentx.h"
 #include "mixer_scheduler.h"
+#include "tasks/mixer_task.h"
 
 bool mixerSchedulerWaitForTrigger(uint8_t timeoutMs)
 {
