+++ radio/src/spacemouse.cpp
@@ -20,7 +20,8 @@
  */
 
 #include "opentx.h"
-#include <ctype.h>
+// #include <ctype.h>
+#include "tasks/mixer_task.h"
 
 #if !defined(SIMU)
   #include <FreeRTOS/include/FreeRTOS.h>
@@ -62,7 +63,7 @@ static StaticTimer_t spacemouseTimerBuffer;
 
 static void spacemouseTimerCb(TimerHandle_t xTimer)
 {
-  if (!s_pulses_paused) {
+  if (!mixerTaskRunning()) {
     DEBUG_TIMER_START(debugTimerSpacemouseWakeup);
     spacemousePoll();
     spacemouseWakeup();
